# specs/001-competency-pdi-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Competency + Seniority (Foundations)
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
  /api/admin/job-titles:
    get:
      summary: List global job titles
      responses:
        '200':
          description: OK
    post:
      summary: Create a global job title (super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, hierarchy_level]
              properties:
                name: { type: string }
                hierarchy_level: { type: integer, minimum: 0, maximum: 3 }
                mission: { type: string }
      responses:
        '201': { description: Created }
  /api/admin/competency-frameworks:
    get:
      summary: List global competency frameworks
      responses:
        '200': { description: OK }
    post:
      summary: Create a new framework version for a job title (super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_title_id, structure]
              properties:
                job_title_id: { type: string, format: uuid }
                structure: { type: object }
      responses:
        '201': { description: Created }
...